### 2019-07-09 #31 @whedon
[PRE REVIEW]: RayTracer.jl: A General Purpose Differentiable Renderer
**Submitting author:** @avik-pal (<a href="http://orcid.org/0000-0002-3938-7375">Avik Pal</a>)
**Repository:** <a href="https://github.com/avik-pal/RayTracer.jl" target ="_blank">https://github.com/avik-pal/RayTracer.jl</a>

Thanks for submitting your paper to JuliaCon @avik-pal. The JuliaCon editor @matbesancon, will work with you on this issue to find a reviewer for your submission before creating the main review issue.

@avik-pal if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](http://bit.ly/2V9GSFq) have already agreed to review for JuliaCon and may be suitable for this submission.

The JuliaCon submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
