### 2019-07-09 #31 @matbesancon
Fixing #27
