### 2019-07-09 #31 @snoeyink
This paper introduces a raytracer built in Julia that allows the Zygote Automatic Differentiation tool to compute gradients of parameters.
One can then optimize parameter values for scene understanding or reconstruction.

The examples shown are toy experiments of obtaining camera or light parameters to match a rendered image. They demonstrate that
the combination works, which is good.

I would like to see some discussion of more interesting parameters, such as determining material properties from images. And more on
limitations: determining geometry seems out of reach because there are so many discrete choices: the problem is not differentiable.
How fast is the gradient computation after AD? (How fast is the AD?) How do these grow with scene complexity?

This paper shows the hallmarks of being written in haste as the package is being developed.  I have many comments and suggestions in pdf uploaded to the package repository.

The abstract discusses training times for ML, which is not in the paper, so here is a rewrite:
In this paper we present RayTracer.jl, a renderer in Julia that is fully differentiable via Zygote.jl. This means that RayTracer not only renders 2D images from 3D scene parameters, but it can be used to optimize for model parameters that generate a target image in a Differentiable Programming (DP) pipeline. We interface our renderer with the deep learning library Flux for use in combination with neural networks. We demonstrate the use of this differentiable renderer in rendering tasks and in solving inverse graphics problems.
