### 2019-07-09 avik-pal/RayTracer.jl#9 @snoeyink
JuliaCon submission comments
Here are some comments on your JuliaCon submission in pdf.

The main things are:
1. Title and abstract mislead me about the paper contents, so I have suggestions. Feel free to use your own rewordings, since I may have misplaced the emphasis.
2. Experiments need enough detail that they can be replicated, esp for BVH and light positioning.
3. Plots in fig 1 & 2 should factor out the expected trend and be single column.
4. Little, but important details: drop the brackets around figure & algorithm numbers (makes them look like paper citations.), and keep the aspect ratio from Fig 4 a,b in c,d.
5. Comments on writing, particularly in consistent use of terms and context first.  See Gopen & Swan's https://cseweb.ucsd.edu/~swanson/papers/science-of-writing.pdf
6. The issue of ray tracing being non-differentiable is mentioned in the first paragraph of the intro, but you never say how that affects what you can attempt.  Good science points out the limitations of a technique.

See attached pdfs
[Comments by page on 10.21105.jcon.00031 (1).pdf](https://github.com/avik-pal/RayTracer.jl/files/3374859/Comments.by.page.on.10.21105.jcon.00031.1.pdf)
[Comments on 10.21105.jcon.00031 (1).pdf](https://github.com/avik-pal/RayTracer.jl/files/3374860/Comments.on.10.21105.jcon.00031.1.pdf)
