### 2019-07-10 avik-pal/RayTracer.jl#9 @avik-pal
@snoeyink Thank you for the detailed review. I am working on an updated version of the paper keeping all the points in mind.

> What is the support for this claim? How does the non-differentiability of model-building decisions limit it?

I am adding a current limitations section to the paper where I shall explain the cases where the RayTracer should not be used and why we did not handle those cases explicitly.

> Why is the number of pixels the measure? I'd expect it to be the number of mesh elements.

The main issue here is to measure the time/memory in a way that it stays consistent with increasing the mesh elements. If we increase the number of elements in the mesh it is not necessary that the time to render it will increase. It will be dependent on the configuration of the mesh (so the dependency would not only be the no of mesh elements but a variety of other factors). Hence, the only way I could consistently increase the complexity by a single parameter was to increase the no of pixels keeping the mesh constant.

> How fast is the gradient computation after AD? (How fast is the AD?) How do these grow with scene complexity?

I will put a comparison between the performance of AD vs Numerical Gradients.
