### 2019-07-10 b0cc92c
Add numerical gradient checks
