+++ paper/paper.tex
@@ -23,37 +23,29 @@
 \maketitle
 
 \begin{abstract}
-With the rising popularity of deep learning techniques, being able to differentiate through arbitrary functions is becoming more and more important. These arbitrary functions contain a lot of implicit knowledge stored in them which when used in combination with a neural network can significantly speed up its training. In this paper we present a fully differentiable renderer, which is designed to be integrated in any Differentiable Programming (DP) pipeline. We interface our renderer with the deep learning library Flux for use in combination with neural networks. We demonstrate the use of this differentiable renderer in complex rendering tasks and show its use in solving inverse graphics problems. 
-
-% \todo{Character count should not exceed 600}
+In this paper we present RayTracer.jl, a renderer in Julia that is fully differentiable using source-to-source Automatic Differentiation (AD). This means that RayTracer not only renders 2D images from 3D scene parameters, but it can be used to optimize for model parameters that generate a target image in a Differentiable Programming (DP) pipeline. We interface our renderer with the deep learning library Flux for use in combination with neural networks. We demonstrate the use of this differentiable renderer in rendering tasks and in solving inverse graphics problems.
 
 \headingtable
 
 \end{abstract}
 
-% \let\thefootnote\relax\footnote{* Equal Contribution}
-
-\section{Introduction}
-\label{intro}
-
-In Computer Graphics, a popular technique for rendering photo realistic images is raytracing. Since ray tracing leverages the properties of the materials of the objects in the scene, a natural extension to the rendering problem would be to extract the exact properties of the materials, lighting, etc. given an image of the scene. However, since ray tracing is non-differentiable at some points and calculating the analytic gradients is a very tedious task, it has been difficult to present a general inverse rendering method. As such there is only one framework in our knowledge, redner\cite{Li:2018:DMC}, which has been able to do so by using analytic gradients.
-
-In this paper, we explore the idea of differentiability through a renderer, by leveraging the AD in Julia\cite{bezanson2017julia}. We present a fully general renderer capable of handling complex scenes and able to differentiate through them. We don't rely on analytic gradients but use source-to-source AD to generate efficient gradient code in the backward pass.
-
 % IMAGES
 
 \begin{figure*}[!htb]
     \centering
-    \includegraphics[width=\textwidth]{images/performance-speed.png}
-    \caption{Performance Gains on using BVH}
-    \label{fig:bvh_perf}
-\end{figure*}
-
-\begin{figure*}[!htb]
-    \centering
-    \includegraphics[width=\textwidth]{images/performance-memory.png}
-    \caption{Memory Allocation comparison between rendering with and without BVH}
-    \label{fig:bvh_mem}
+    \begin{subfigure}[c]{0.45\textwidth}
+        \includegraphics[height=75mm, width=\textwidth]{images/performance-speed.png}
+        \caption{Performance Benchmarks}
+        \label{fig:bvh_perf}
+    \end{subfigure}
+    \hfill
+    \begin{subfigure}[c]{0.45\textwidth}
+        \includegraphics[height=75mm, width=\textwidth]{images/performance-memory.png}
+        \caption{Memory Allocation Benchmarks}
+        \label{fig:bvh_mem}
+    \end{subfigure}
+    \caption{Comparison between scenes rendered with and without BVH}
+    \label{fig:bvh}
 \end{figure*}
 
 \begin{figure*}[!htb]
@@ -75,43 +67,51 @@ In this paper, we explore the idea of differentiability through a renderer, by l
         \includegraphics[width=\textwidth]{images/render/teapot_top.jpg}
         \caption*{}
     \end{subfigure}
-    \caption{Utah Teapot Render from three different views. The camera definition shown in Listing~\protect[\ref{lst:example_render}] can be easily modified to generate all these views.}
[…]
