+++ test/gradcheck.jl
@@ -13,23 +13,35 @@ end
 
 screen_size = (w = 400, h = 300)
 
-light = PointLight(Vec3(1.0f0), 1000.0f0, Vec3(0.15f0, 0.5f0, -110.5f0))
+light = PointLight(
+    Vec3(1.0),
+    1000.0,
+    Vec3(0.15, 0.5, -110.5)
+)
 
-eye_pos = Vec3(0.0f0, 0.0f0, -5.0f0)
+eye_pos = Vec3(0.0, 0.0, -5.0)
 
-cam = Camera(eye_pos, Vec3(0.0f0), Vec3(0.0f0, 1.0f0, 0.0f0), 45.0f0, 1.0f0, screen_size...)
+cam = Camera(eye_pos, Vec3(0.0), Vec3(0.0, 1.0, 0.0), 45.0, 1.0, screen_size...)
 
 origin, direction = get_primary_rays(cam);
 
 @testset "Triangle" begin
     scene = [
-        Triangle(Vec3(-1.7f0, 1.0f0, 0.0f0), Vec3(1.0f0, 1.0f0, 0.0f0), Vec3(1.0f0, -1.0f0, 0.0f0),
-                 Material())
+        Triangle(Vec3(-1.7, 1.0, 0.0), Vec3(1.0, 1.0, 0.0), Vec3(1.0, -1.0, 0.0),
+                 Material(color_ambient = Vec3(1.0),
+                          color_diffuse = Vec3(1.0),
+                          color_specular = Vec3(1.0),
+                          specular_exponent = 50.0,
+                          reflection = 1.0))
     ]
 
     scene_new = [
-        Triangle(Vec3(-1.9f0, 1.3f0, 0.1f0), Vec3(1.2f0, 1.1f0, 0.3f0), Vec3(0.8f0, -1.2f0, -0.15f0),
-                 Material())
+        Triangle(Vec3(-1.9, 1.3, 0.1), Vec3(1.2, 1.1, 0.3), Vec3(0.8, -1.2, -0.15),
+                 Material(color_ambient = Vec3(1.0),
+                          color_diffuse = Vec3(1.0),
+                          color_specular = Vec3(1.0),
+                          specular_exponent = 50.0,
+                          reflection = 1.0))
     ]
     
     color = raytrace(origin, direction, scene, light, eye_pos, 0)
@@ -38,17 +50,16 @@ origin, direction = get_primary_rays(cam);
 
     numerical_grads = get_params(numderiv(x -> loss_fn([x], color), scene_new[1]))
 
-    # Ignore the Material Gradients
-    @test isapprox(numerical_grads[1:end-4], zygote_grads[1:end-4], rtol = 1.0e-1)
+    @test isapprox(numerical_grads, zygote_grads, rtol = 1.0e-3)
 end
     
 @testset "Sphere" begin
     scene = [
-        Sphere(Vec3(-1.7f0, 1.0f0, 0.0f0), [0.6f0], Material())
+        Sphere(Vec3(-1.7, 1.0, 0.0), [0.6], Material())
     ]
 
     scene_new = [
-        Sphere(Vec3(-1.9f0, 1.3f0, 0.1f0), [0.9f0], Material())
+        Sphere(Vec3(-1.9, 1.3, 0.1), [0.9], Material())
     ]
     
     color = raytrace(origin, direction, scene, light, eye_pos, 0)
@@ -57,6 +68,5 @@ end
 
     numerical_grads = get_params(numderiv(x -> loss_fn([x], color), scene_new[1]))
 
-    # Ignore the Material Gradients
-    @test isapprox(numerical_grads[1:end-5], zygote_grads[1:end-5], rtol = 1.0e-1)
+    @test isapprox(numerical_grads, zygote_grads, rtol = 1.0e-3)
 end
