+++ paper/header.tex
@@ -1,9 +1,10 @@
 % **************GENERATED FILE, DO NOT EDIT**************
 
-\title{RayTracer.jl: A General Purpose Differentiable Renderer}
+\title{RayTracer.jl: A Differentiable Renderer that supports Parameter Optimization for Scene Reconstruction}
 
 \author[1]{Avik Pal}
-\affil[1]{Indian Institute of Technology, Kanpur}
+\affil[1]{Indian Institute of Technology Kanpur}
 
-\keywords{Julia, Differentiable Programming, Deep Learning, Automatic Differentiation, Inverse Graphics}
+\keywords{Julia, Differentiable Programming, Automatic Differentiation, Inverse Graphics, Differentiable Rendering}
+
 
