### 2019-07-11 6580a6a
Update the paper and the toml files
