### 2019-07-11 5368622
Update paper as per review
