### 2019-07-15 #31 @dourouc05
I would be available, starting from Wednesday! (For the process, shall I put the review here? Or is there a more detailed guide?)
