### 2019-07-15 #31 @matbesancon
> For the process, shall I put the review here? Or is there a more detailed guide?

Yes there is a guide [here](https://proceedings.juliacon.org/guide/reviewers/), the review issue will be opened soon with a checklist
