### 2019-07-15 #37 @whedon
[REVIEW]: RayTracer.jl: A General Purpose Differentiable Renderer
**Submitting author:** @avik-pal (<a href="http://orcid.org/0000-0002-3938-7375">Avik Pal</a>)
**Repository:** <a href="https://github.com/avik-pal/RayTracer.jl" target ="_blank">https://github.com/avik-pal/RayTracer.jl</a>
