### 2019-07-17 5a9038e
Update paper.tex
