### 2019-07-17 avik-pal/RayTracer.jl#9 @dourouc05
A few other comments (I build upon what @snoeyink said and thus not repeating too much). I really like the approach and your use cases! That's really something that more people should know about, and more software like this is a great step in that direction.

Major comments:

* Did you perform any hyperparameter optimisation? (Like learning rate, parameters for Adam.) If so, how?
* Overall, your submission felt a bit lacking: no performance figures, no real evaluation of the optimisation algorithms (Adam is not the only one) or of the hyperparameters; also, you don't tell a word about Duckietown, although I think this is a very practical use case.
* "We interface our renderer with the deep learning library Flux for use in combination with neural networks": what combinations are you making with neural networks? This is the only mention of neural networks in the whole submission I can find.
* You say that automatic differentiation is not so available in languages like C++, but there are packages like CppAD (https://github.com/coin-or/CppAD) or CasADi (https://web.casadi.org/): you should mention them and discuss why they are not applicable.
* How do you compare to redner? You just mention its existence, but do not compare yourself to it. Your package doesn't really have to shine, but that would really be useful information to user to decide which library they should use/invest on (functionalities, performance, ease of extension, etc.).

Minor ones:

* Figure 1 is not very legible, you seem to have stretched the image to make it fit in the page. (The same is true, albeit to a lesser extent, for Figures 3 and 4. For Figure 4, try to have all plot titles have the same size.)
* "For these measurements we use a mesh with 137 centered": a word is missing (137 what?).
* In Algorithm 1, I don't understand your syntax "gs = gradient(params) do". Are you just defining a variable `gs`? If so, why the `do-end`?
* I would add more context to the abstract, to briefly explain a potential use case (within the limits of JuliaCon).

About the software:

* The constructor for `Camera` should use keyword arguments, because as is the code is hard to read (like in your Listing 1): there are too many arguments for position to be enough. The same goes for the function `raytrace`. It wouldn't hurt for PointLight, but it feels less necessary.
* I don't think the package name is well-suited: you might want to have other ray-tracers in Julia, with other trade-offs (better performance but lack of differentiability, for instance).
* For your test suite, I think that a test like https://github.com/avik-pal/RayTracer.jl/blob/master/test/mesh_render.jl could be improved upon a lot: why don't you check the return values? You could have reference images and compare them to the results you get from the code. This is slightly brittle, but that's often the case with integration tests (and I've seen this approach for GUI toolkits). You could remain at the unit test by ensuring that the correct graph is built before being passed to AD, but that would be more difficult to write and, in my opinion, less useful. Otherwise, you really should do a lot more unit testing: testing small bits of your code, each and every public function, so that you can easily spot which function stopped working.
* I feel that your documentation is not polished enough: it is quite good for a prototype, but not enough for software to be published and used by outside people. For instance, in https://avik-pal.github.io/RayTracer.jl/dev/getting_started/rendering/#A-Step-By-Step-Introduction-to-Rendering-1, you give a lot of code, but you don't really explain the parameters you give to each function (you should also have links to the "reference" part of the documentation).
* In your submission, you speak about Optim.jl, but I did not see anything related to that package in your code: is there a reason for this?
* You don't really explain why you need Julia 1.1 (might be worth a small note, because Julia 1.0 is the long-term version and thus may be very relevant for some people).
