+++ src/gradients/numerical.jl
@@ -39,6 +39,43 @@ end
 
 Get the parameters from a struct that can be tuned. The output is in
 the form of an array.
+
+### Example:
+
+```jldoctest
+julia> using RayTracer;
+
+julia> scene = Triangle(Vec3(-1.9f0, 1.6f0, 1.0f0), Vec3(1.0f0, 1.0f0, 0.0f0), Vec3(-0.5f0, -1.0f0, 0.0f0),
+                        Material())
+Triangle Object:
+    Vertex 1 - x = -1.9, y = 1.6, z = 1.0
+    Vertex 2 - x = 1.0, y = 1.0, z = 0.0
+    Vertex 3 - x = -0.5, y = -1.0, z = 0.0
+    Material{Array{Float32,1},Array{Float32,1},Nothing,Nothing,Nothing,Nothing}(x = 1.0, y = 1.0, z = 1.0, x = 1.0, y = 1.0, z = 1.0, x = 1.0, y = 1.0, z = 1.0, Float32[50.0], Float32[0.5], nothing, nothing, nothing, nothing)
+
+julia> RayTracer.get_params(scene)
+20-element Array{Float32,1}:
+ -1.9
+  1.6
+  1.0
+  1.0
+  1.0
+  0.0
+ -0.5
+ -1.0
+  0.0
+  1.0
+  1.0
+  1.0
+  1.0
+  1.0
+  1.0
+  1.0
+  1.0
+  1.0
+ 50.0
+  0.5
+```
 """
 get_params(x::T, typehelp = Float32) where {T<:AbstractArray} = x
 
@@ -57,25 +94,50 @@ Sets the tunable parameters of the struct `x`. The index of the last element
 set into the struct is returned as output. This may be used to confirm that
 the size of the input array was as expected.
 
-### Example
+### Example:
 
-```julia
-julia> scene = Triangle(Vec3(-1.9, 1.3, 0.1), Vec3(1.2, 1.1, 0.3), Vec3(0.8, -1.2, -0.15),
-                    color = rgb(1.0, 1.0, 1.0), reflection = 0.5)
-Triangle{Array{Float64,1}}(Vec3{Array{Float64,1}}([-1.9], [1.3], [0.1]), Vec3{Array{Float64,1}}([1.2], [1.1], [0.3]), Vec3{Array{Float64,1}}([0.8], [-1.2], [-0.15]), RayTracer.Material{RayTracer.PlainColor,Float64}(RayTracer.PlainColor(Vec3{Array{Float64,1}}([1.0], [1.0], [1.0])), 0.5))
+```jldoctest
+julia> using RayTracer;
 
-julia> x = rand(13)
-13-element Array{Float64,1}:
- 0.39019817669623835
- 0.940810689314205
- .
- .
- .
- 0.5590307650917048
- 0.7551647340674075
+julia> scene = Triangle(Vec3(-1.9f0, 1.6f0, 1.0f0), Vec3(1.0f0, 1.0f0, 0.0f0), Vec3(-0.5f0, -1.0f0, 0.0f0),
+                        Material())
+Triangle Object:
+    Vertex 1 - x = -1.9, y = 1.6, z = 1.0
+    Vertex 2 - x = 1.0, y = 1.0, z = 0.0
+    Vertex 3 - x = -0.5, y = -1.0, z = 0.0
[…]
