+++ docs/make.jl
@@ -1,14 +1,28 @@
 using Documenter, RayTracer
 
-makedocs(modules=[RayTracer],
-         doctest = false,
-         sitename = "RayTracer",
-         authors = "Avik Pal",
-         format = Documenter.HTML(prettyurls = true,
-                                  assets = ["assets/documenter.css"]),
-         pages = ["Home" => "index.md",
-                  "Getting Started" => [
-                      "Introduction to Rendering" => "getting_started/teapot_rendering.md"],
-                  "API Documentation" => "api.md"])
+makedocs(
+    modules = [RayTracer],
+    doctest = true,
+    sitename = "RayTracer",
+    authors = "Avik Pal",
+    format = Documenter.HTML(
+        prettyurls = get(ENV, "CI", nothing) == "true",
+    ),
+    pages = [
+        "Home"              => "index.md",
+        "Getting Started"   => [
+            "Introduction to Rendering" => "getting_started/teapot_rendering.md",
+        ],
+        "API Documentation" => [
+            "General Utilities"   => "api/utilities.md",
+            "Differentiation"     => "api/differentiation.md",
+            "Scene Configuration" => "api/scene.md",
+            "Renderers"           => "api/renderers.md",
+            "Optimization"        => "api/optimization.md",
+        ],
+    ],
+)
 
-deploydocs(repo = "github.com/avik-pal/RayTracer.jl.git")
+deploydocs(
+    repo = "github.com/avik-pal/RayTracer.jl.git"
+)
