+++ src/optimize.jl
@@ -20,11 +20,11 @@ type conversion of `x` which can significantly slow doen the raytracer.
 ### Example:
 
 ```julia
-opt = ADAM()
+julia> opt = ADAM()
 
-gs = gradient(loss_function, θ)
+julia> gs = gradient(loss_function, θ)
 
-update!(opt, θ, gs[1])
+julia> update!(opt, θ, gs[1])
 ```
 """
 function update!(opt, x::AbstractArray, Δ::AbstractArray)
@@ -33,11 +33,6 @@ function update!(opt, x::AbstractArray, Δ::AbstractArray)
 end
 
 update!(opt, x::T, Δ::T) where {T<:Real} = (update!(opt, [x], [Δ]))[]
-
-# This makes sure we donot end up optimizing the value of the material.
-# We cannot do this update in a stable manner for now. So it is wise
-# to just avoid it for now.
-# update!(opt, x::Material, Δ::Material) = x
 
 update!(opt, x::FixedParams, Δ::FixedParams) = x 
 
