+++ docs/src/api/optimization.md
@@ -1,0 +1,19 @@
+```@meta
+CurrentModule = RayTracer
+```
+
+# Optimization
+
+RayTracer.jl can be used for inverse rendering problems. See `examples/` for more details on
+these forms of problems.
+
+```@index
+Pages = ["api/optimization.md"]
+```
+
+## Documentation
+
+```@autodocs
+Modules = [RayTracer]
+Pages = ["optimize.jl"]
+```

[…]
