+++ docs/src/api.md
@@ -1,16 +1,0 @@
-# API Documentation
-
-```@meta
-CurrentModule = RayTracer
-```
-
-
-## Renderers
-
-```@autodocs
-Modules = [RayTracer]
-Pages = ["blinnphong.jl"]
-Order = [:type,
-         :macro,
-         :function]
-```
