### 2019-07-19 2aebaeb
Final documentation additions
