+++ .travis.yml
@@ -19,11 +19,12 @@ jobs:
       julia: 1.1
       os: linux
       script:
-        - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd()));
-                                               Pkg.instantiate()'
-        - julia --project=docs/ docs/make.jl
+        - julia --project=docs/ --color=yes -e 'using Pkg;
+                                                Pkg.develop(PackageSpec(path=pwd()));
+                                                Pkg.instantiate()'
+        - julia --project=docs/ --color=yes --compile=min docs/make.jl
       after_success: skip
       
 script:
   - julia --color=yes -e 'using Pkg; Pkg.activate(); Pkg.instantiate(); Pkg.test(coverage=true)'
-  - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Codecov.submit(Codecov.process_folder())'
+  - julia --color=yes -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Codecov.submit(Codecov.process_folder())'
