### 2019-07-19 4780b82
Improve the documentation css and README
