+++ examples/teapot_rendering.jl
@@ -7,7 +7,8 @@
 # Run this code in your terminal to get the file:
 # `wget https://raw.githubusercontent.com/McNopper/OpenGL/master/Binaries/teapot.obj`
 
-using RayTracer, Images
+# If you are using REPL mode you need the `ImageView.jl` package
+using RayTracer, Images #, ImageView
 
 # ## General Attributes of the Scene
 #
@@ -38,13 +39,14 @@ function generate_render_and_save(cam, light, filename)
     img = get_image(color, screen_size...)
 
     #src # Display the image
+    #src # For REPL mode change this to `imshow(img)`
     display(img)
 
     #src # Save the generated image
     save(filename, img)
 end
 
-# ## Understanding the `Light` and `Camera` API
+# ## Understanding the [`Light`](@ref) and [`Camera`](@ref) API
 #
 # ### DistantLight
 #
