### 2019-07-19 2fcf652
Add usage example with Optim
