+++ paper/paper.tex
@@ -376,7 +376,7 @@ light_target = PointLight(
                    language = Julia]
 light_guess = PointLight(
     Vec3(1.0f0, 1.0f0, 1.0f0),
-    10000.0f0,
+    1.0f0,
     Vec3(-1.0f0, -10.0f0, -50.0f0)
 )
 \end{lstlisting}
