+++ examples/teapot_rendering.jl
@@ -46,7 +46,7 @@ function generate_render_and_save(cam, light, filename)
     save(filename, img)
 end
 
-# ## Understanding the [`Light`](@ref) and [`Camera`](@ref) API
+# ## Understanding the Light and Camera API
 #
 # ### DistantLight
 #
