### 2019-07-20 avik-pal/RayTracer.jl#9 @avik-pal
@dourouc05 Thank you for your feedback.

## Software Concerns

I have tried to address the software issues you mentioned in #13
1. Now `raytrace`, `Camera` and `Light` have keyword arguments. But the original API also exists.
2. All the functions are now documented with some detailed getting started tutorials.
3. I have added a tutorial on how to use Optim for scene reconstruction
4. Added a note on why julia 1.0 cannot be used

I will deal with the remaining software Issues like the test suite in the next few days.
