### 2019-07-20 157930b
Changes suggested in code review
