### 2019-07-21 avik-pal/RayTracer.jl#9 @avik-pal
I will get back to the other comments once JuliaCon is over.
