### 2019-07-21 avik-pal/RayTracer.jl#9 @dourouc05
No problem, you already did a good job :+1:!
