### 2019-07-22 avik-pal/RayTracer.jl#9 @snoeyink
I too appreciate the changes that you made and think they are a definite improvement.

Here are some typos that you missed. The experiment that generates Fig 3 could use more detail to make it repeatable.  [Summary of Comments on 10.21105.jcon.00037.pdf](https://github.com/avik-pal/RayTracer.jl/files/3418217/Summary.of.Comments.on.10.21105.jcon.00037.pdf)
[10.21105.jcon.00037.pdf](https://github.com/avik-pal/RayTracer.jl/files/3418218/10.21105.jcon.00037.pdf)

Enjoy JuliaCon.
