### 2019-07-23 #37 @snoeyink
added comments in [avik-pal/RayTracer.jl#9](https://github.com/avik-pal/RayTracer.jl/issues/9) on latest version, which is an improvement.
