### 2019-07-24 57f974a
Update travis script
