### 2019-07-29 avik-pal/RayTracer.jl#9 @avik-pal
Just a small note, even though I might have mentioned some urls in the paper which refer to the master branch, they are currently in the `review_updates`. In future I will list them in this issue separately.
[https://github.com/avik-pal/RayTracer.jl/examples](https://github.com/avik-pal/RayTracer.jl/examples) --> [https://github.com/avik-pal/RayTracer.jl/tree/review_updates/examples](https://github.com/avik-pal/RayTracer.jl/tree/review_updates/examples)
