### 2019-07-30 avik-pal/RayTracer.jl#9 @avik-pal
> In Algorithm 1, I don't understand your syntax "gs = gradient(params) do". Are you just defining a variable gs? If so, why the do-end?

The `do-end` syntax simply transforms the function enclosed in it to the following form:
```
gs = gradient(params) do
     l1 = operation1(params)
     return operation2(l1)
end
```

```
function myfunction(s)
     l1 = operation1(s)
     return operation2(l1)
end
gs = gradient(myfunction, params)
```
