### 2019-07-30 avik-pal/RayTracer.jl#9 @dourouc05
I think it would be best to avoid this syntax in the article, as this syntax is quite specific to Julia.
