### 2019-08-01 63814f5
Update the paper as per the review comments
