+++ paper/paper.tex
@@ -23,30 +23,13 @@
 \maketitle
 
 \begin{abstract}
-In this paper, we present RayTracer.jl, a renderer in Julia that is fully differentiable using source-to-source Automatic Differentiation (AD). This means that RayTracer not only renders 2D images from 3D scene parameters, but it can be used to optimize for model parameters that generate a target image in a Differentiable Programming (DP) pipeline. We interface our renderer with the deep learning library Flux for use in combination with neural networks. We demonstrate the use of this differentiable renderer in rendering tasks and in solving inverse graphics problems.
+In this paper, we present RayTracer.jl, a renderer in Julia that is fully differentiable using source-to-source Automatic Differentiation (AD). This means that RayTracer not only renders 2D images from 3D scene parameters, but it can be used to optimize for model parameters that generate a target image in a Differentiable Programming (DP) pipeline. Our renderer is a complete general purpose renderer which means that unlike most previous work, we don't make any changes to the renderer to make it differentiable. Additionally we interface our renderer with the deep learning library Flux for use in combination with neural networks. In this paper, we also demonstrate the use of this differentiable renderer in rendering tasks and in solving inverse graphics problems using gradients.
 
 \headingtable
 
 \end{abstract}
 
 % IMAGES
-
-\begin{figure*}[!htb]
-    \centering
-    \begin{subfigure}[c]{0.45\textwidth}
-        \includegraphics[height=75mm, width=\textwidth]{images/performance-speed.png}
-        \caption{Performance Benchmarks}
-        \label{fig:bvh_perf}
-    \end{subfigure}
-    \hfill
-    \begin{subfigure}[c]{0.45\textwidth}
-        \includegraphics[height=75mm, width=\textwidth]{images/performance-memory.png}
-        \caption{Memory Allocation Benchmarks}
-        \label{fig:bvh_mem}
-    \end{subfigure}
-    \caption{Comparison between scenes rendered with and without BVH}
-    \label{fig:bvh}
-\end{figure*}
 
 \begin{figure*}[!htb]
     \centering
@@ -74,44 +57,30 @@ In this paper, we present RayTracer.jl, a renderer in Julia that is fully differ
 \section{Introduction}
 \label{intro}
 
-Rendering is a technique of generating photorealistic or non-photorealistic 2D projections from 3D objects. As such there are several algorithms for rendering complex scenes. One of the most popular techniques for photorealistic rendering is ray tracing. However, for real-time rendering algorithms like rasterization is used.
+Rendering is a technique of generating photo-realistic or non-photo-realistic 2D projections from 3D objects. As such there are several algorithms for rendering complex scenes. One of the most popular techniques for photo-realistic rendering is ray tracing. However, for real-time rendering algorithms like rasterization are used.
 
-Ray Tracing is a technique in computer graphics for rendering 3D graphics with complex light interactions. In this technique, rays are traced backward from the eye/camera to the light source(s). The ray can undergo reflection and refraction due to interactions with the objects in its path. This technique, however, is very computationally expensive and hence difficult to do in real-time. But since ray tracing leverages the properties of the materials of the objects in the scene, a natural extension to the rendering problem would be to extract the exact properties of the materials, lighting, etc. given an image of the scene. Unfortunately, ray tracing is non-differentiable at some points and calculating the analytic gradients is a very tedious task. This has made it a very difficult task to present a general inverse rendering method. As such, there is only one framework in our knowledge, redner~\cite{Li:2018:DMC}, which has been able to do so by using analytic gradients.
[…]
