+++ paper/ref.bib
@@ -39,12 +39,13 @@
 }
 
 @misc{jax,
-  author = {},
-  title = {Jax},
+  title = {{JAX}: {A}utograd and {XLA}},
+  author = {Matt Johnson and
+            Roy Frostig and
+            Dougal Maclaurin and
+            Chris Leary},
   year = {2018},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  howpublished = {\url{https://github.com/google/jax}}
+  howpublished = {\url{https://github.com/google/jax}},
 }
 
 @article{Flux.jl-2018,
@@ -133,4 +134,13 @@
   timestamp = {Mon, 27 May 2019 13:15:00 +0200},
   biburl    = {https://dblp.org/rec/bib/journals/corr/abs-1905-00641},
   bibsource = {dblp computer science bibliography, https://dblp.org}
+}
+
+@article{casadi,
+  Author = {Joel A E Andersson and Joris Gillis and Greg Horn
+            and James B Rawlings and Moritz Diehl},
+  Title = {{CasADi} -- {A} software framework for nonlinear optimization
+           and optimal control},
+  Journal = {Mathematical Programming Computation},
+  Year = {In Press, 2018},
 }
