### 2019-08-01 e6ce0fa
BVH doc fixes
