### 2019-08-01 avik-pal/RayTracer.jl#9 @snoeyink
I am completely fine with `do` syntax in a JuliaCon proceedings paper, although @dourouc05's comment suggests that it would be good to briefly explain it.

Some comments on the latest revision, which continues to improve on the original submission.

Figure 3 and the revised Fig 4 are nice additions.
I did have to read the whole of section 5.2 before I can understand the context for the first paragraph of that section -- perhaps you can start with the question that you are trying to answer, rather than a statement "we compare the use".  (The best writers "[show, don't tell.](https://jerryjenkins.com/show-dont-tell/)"  Think about the reader's perspective, and get them to ask the question, "is AD better than numerical differentiation?"  just before you provide evidence for the answer.)

Section 2, I like:
> Differentiation allows such renderers to make use of gradients to learn the inverse mapping from an image to its parameter space.

But I think the rest of the paragraph needs evidence in this paper if that is going to be a prominent claim.  The devil is in the details, so a citation to [3] is wishful thinking or complete bullshit.

Fig 2: I strongly suspect is that what the data shows is that the BVH takes half the time (and half the memory allocations, which I suspect is related).  You can't see that in at the left, but only at the right of each plot, since the plots are dominated by the exponential growth in number of pixels/parameters.  Please try plotting time per pixel/parameter (or sqrt(pixel or parameter)) rather than total time to see if this reveals if the multiplier is consistent at 1/2.

The review checklist suggests that DOI's should be included for references that have them.
