+++ paper/paper.tex
@@ -67,10 +67,7 @@ In this paper, we explore the idea of differentiability through a renderer, by l
                                                                                                                               
 \section{Differentiable Ray Tracing}
 
-There are several photo-realistic renderers available which contain a vast amount of implicit knowledge. Differentiation allows such renderers to make use of gradients to learn the inverse mapping from an image to its parameter space. We can then use this knowledge in combination with any machine learning / deep learning models to train them in a fast and efficient manner. Experiments in~\cite{retinaface} demonstrate the effectiveness of incorporating differentiable renderers in deep learning pipelines by achieving State of the Art results in WIDERFace Hard dataset.
-
-However, as usual, it is difficult to compute efficient derivatives from a production-ready renderer, typically written in a performance language like C++. This provides the primary motivation towards the development of RayTracer.jl. We develop an entire general purpose ray tracer in a high-level numerical computation language. The presence of strong automatic differentiation libraries like Zygote.jl make it trivial to compute efficient derivatives from the renderer.We present the performance gains we get on using Zygote as compared to Central Differencing in Section~\ref{sec:finite_diff}.
-
+There are several photo-realistic renderers available which contain a vast amount of implicit knowledge. Differentiation allows such renderers to make use of gradients to learn the inverse mapping from an image to its parameter space. However as usual, it is difficult to compute efficient derivatives from a production-ready renderer, typically written in a performance language like C++. This provides the primary motivation towards the development of RayTracer.jl. We develop an entire general purpose ray tracer in a high-level numerical computation language. The presence of strong automatic differentiation libraries like Zygote.jl make it trivial to compute efficient derivatives from the renderer. We present the performance gains we get on using Zygote as compared to Central Differencing in Section~\ref{sec:finite_diff}.
 
 RayTracer.jl~\cite{RayTracer.jl} is a package for Differentiable Ray Tracing written to solve this particular issue. It relies heavily on the source-to-source automatic differentiation package, Zygote, for computing gradients with respect to arbitrary scene parameters. This package allows the user to configure the location of objects, lights and a camera in the scene. This scene is then interpreted by the renderer to generate the image. RayTracer.jl is naturally interfaced with the deep learning library Flux \cite{Flux.jl-2018}, due to the common AD backend, for use in more complex differentiable pipelines.
 
@@ -198,9 +195,9 @@ We present the benefits of using BVH in Figure~\ref{fig:bvh}. We are able to get
 \subsection{Comparison with Finite Differencing}
 \label{sec:finite_diff}
 
-In this section, we compare the use of our source-to-source AD backend with Central Finite Differencing, that is built into the package for gradient testing purposes. For comparing the two differentiation techniques, we run five independent trails and present the mean runtimes (with the standard deviation) in Figure~\ref{fig:ad_better}.
+In this section, we demonstrate the performance gain of using source-to-source AD backend versus Central Finite Differencing, that is built into the package for gradient testing purposes. Through this experiment we show that our method is exponentially better than numerical differentiation.
 
-We fix the position of the camera and light and randomly generate the triangles present in the scene. Every new triangle added to the scene, adds 20 additional parameters with respect to which we must compute the derivatives. We use the mean squared loss function to compute the scalar loss and backpropagate.
+For comparing the two differentiation techniques, we run five independent trails and present the mean runtimes (with the standard deviation) in Figure~\ref{fig:ad_better}. We fix the position of the camera and light and randomly generate the triangles present in the scene. Every new triangle added to the scene, adds 20 additional parameters with respect to which we must compute the derivatives. We use the mean squared loss function to compute the scalar loss and backpropagate.
 
[…]
