+++ paper/ref.bib
@@ -1,12 +1,13 @@
 @article{bezanson2017julia,
-  title={Julia: A fresh approach to numerical computing},
-  author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
-  journal={SIAM review},
-  volume={59},
-  number={1},
-  pages={65--98},
-  year={2017},
-  publisher={SIAM}
+  title = {Julia: A fresh approach to numerical computing},
+  author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
+  journal = {SIAM review},
+  volume = {59},
+  number = {1},
+  pages = {65--98},
+  year = {2017},
+  publisher = {SIAM},
+  doi = {10.1137/141000671}
 }
 
 @article{Li:2018:DMC,
@@ -16,7 +17,8 @@
     volume = {37},
     number = {6},
     pages = {222:1--222:11},
-    year = {2018}
+    year = {2018},
+    doi = {10.1145/3272127.3275109}
 }
 
 @article{kingma2014adam,
@@ -143,4 +145,5 @@
            and optimal control},
   Journal = {Mathematical Programming Computation},
   Year = {In Press, 2018},
+  doi = {10.1007/s12532-018-0139-4},
 }
