### 2019-08-12 avik-pal/RayTracer.jl#9 @avik-pal
I have added the DOI's in the latest revision
