### 2019-08-28 avik-pal/RayTracer.jl#9 @dourouc05
Some proofreading would help (most notably with commas: you use very few of them, which makes the text rather hard to read). Two typos in Section 5.5: "The converge of the model […] paramters" -> convergence and parameters.

Figure 3b: the y-axis is not clear (acceleration factor?). Also, how good were the gradients computed by finite differences? Maybe you could give some insight into this method? Also, how do you choose the step to apply the formula (i.e. the h in (f(x-h) - f(x+h)/2h)?

A bit of nit-picking: in Algo. 1, you never define the value of `iter`, nor do you update it. Similarly, `converged` is never set to `false`, and there is no explicit initial value for `params`.
