### 2019-08-31 28fe8e0
Update minor typos in the paper
