+++ paper/paper.tex
@@ -23,7 +23,7 @@
 \maketitle
 
 \begin{abstract}
-In this paper, we present RayTracer.jl, a renderer in Julia that is fully differentiable using source-to-source Automatic Differentiation (AD). This means that RayTracer not only renders 2D images from 3D scene parameters, but it can be used to optimize for model parameters that generate a target image in a Differentiable Programming (DP) pipeline. Our renderer is a complete general purpose renderer which means that unlike most previous work, we don't make any changes to the renderer to make it differentiable. Additionally we interface our renderer with the deep learning library Flux for use in combination with neural networks. In this paper, we also demonstrate the use of this differentiable renderer in rendering tasks and in solving inverse graphics problems using gradients.
+In this paper, we present RayTracer.jl, a renderer in Julia that is fully differentiable using source-to-source Automatic Differentiation (AD). This means that RayTracer not only renders 2D images from 3D scene parameters, but it can be used to optimize for model parameters that generate a target image in a Differentiable Programming (DP) pipeline. Our renderer is a complete general purpose renderer, which means that unlike most previous work, we do not make any changes to the renderer to make it differentiable. Additionally, we interface our renderer with the deep learning framework Flux for use in combination with neural networks. In this paper, we also demonstrate the use of this differentiable renderer in rendering tasks and in solving inverse graphics problems using gradients.
 
 \headingtable
 
@@ -57,30 +57,34 @@ In this paper, we present RayTracer.jl, a renderer in Julia that is fully differ
 \section{Introduction}
 \label{intro}
 
-Rendering is a technique of generating photo-realistic or non-photo-realistic 2D projections from 3D objects. As such there are several algorithms for rendering complex scenes. One of the most popular techniques for photo-realistic rendering is ray tracing. However, for real-time rendering algorithms like rasterization are used.
+Rendering is a technique of generating photo-realistic or non-photo-realistic 2D projections from 3D objects. As such, there are several algorithms for rendering complex scenes. One of the most popular techniques for photo-realistic rendering is ray tracing~\cite{Appel:1968:TSM:1468075.1468082}. For real-time rendering we use techniques like rasterization~\cite{Pineda:1988:PAP:54852.378457} are used.
 
-Ray Tracing is a technique in computer graphics for rendering 3D graphics with complex light interactions. In this technique, rays are traced backward from the eye/camera to the light source(s). The ray can undergo reflection and refraction due to interactions with the objects in its path. This technique, however, is very computationally expensive and hence difficult to do in real-time. But since ray tracing leverages the properties of the materials of the objects in the scene, a natural extension to the rendering problem would be to extract the exact properties of the materials, lighting, etc. given an image of the scene. Calculating analytic gradients for every single parameter of the scene is a very tedious process and prone to errors. This has made it a difficult task to present a general gradient based inverse rendering method. As such, there is only one framework in our knowledge, redner~\cite{Li:2018:DMC}, which has been able to do so by using analytic gradients. However, we bypass this problem by using AD. Another major limitation that exists in case of differentiability is the presence of geometric discontinuities, like the edge between two triangles. This makes the renderer non-differentiable at those points. Unfortunately this is still one of the limitations (Section~\ref{sec:limitation}) of RayTracer.
+Ray Tracing is a technique in computer graphics for rendering 3D graphics with complex light interactions. In this technique, rays are traced backward from the eye/camera to the light source(s). The ray can undergo reflection and refraction due to interactions with the objects in its path. This technique, however, is very computationally expensive and hence difficult to do in real-time. 
 
[…]
