+++ paper/ref.bib
@@ -1,13 +1,12 @@
 @article{bezanson2017julia,
-  title = {Julia: A fresh approach to numerical computing},
-  author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
-  journal = {SIAM review},
-  volume = {59},
-  number = {1},
-  pages = {65--98},
-  year = {2017},
-  publisher = {SIAM},
-  doi = {10.1137/141000671}
+  title={Julia: A fresh approach to numerical computing},
+  author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
+  journal={SIAM review},
+  volume={59},
+  number={1},
+  pages={65--98},
+  year={2017},
+  publisher={SIAM}
 }
 
 @article{Li:2018:DMC,
@@ -17,8 +16,7 @@
     volume = {37},
     number = {6},
     pages = {222:1--222:11},
-    year = {2018},
-    doi = {10.1145/3272127.3275109}
+    year = {2018}
 }
 
 @article{kingma2014adam,
@@ -145,5 +143,54 @@
            and optimal control},
   Journal = {Mathematical Programming Computation},
   Year = {In Press, 2018},
-  doi = {10.1007/s12532-018-0139-4},
+}
+
+@inproceedings{Appel:1968:TSM:1468075.1468082,
+ author = {Appel, Arthur},
+ title = {{Some Techniques for Shading Machine Renderings of Solids}},
+ booktitle = {Proceedings of the April 30--May 2, 1968, Spring Joint Computer Conference},
+ series = {AFIPS '68 (Spring)},
+ year = {1968},
+ location = {Atlantic City, New Jersey},
+ pages = {37--45},
+ numpages = {9},
+ url = {http://doi.acm.org/10.1145/1468075.1468082},
+ doi = {10.1145/1468075.1468082},
+ acmid = {1468082},
+ publisher = {ACM},
+ address = {New York, NY, USA},
+} 
+
+@inproceedings{Pineda:1988:PAP:54852.378457,
+ author = {Pineda, Juan},
+ title = {{A Parallel Algorithm for Polygon Rasterization}},
+ booktitle = {Proceedings of the 15th Annual Conference on Computer Graphics and Interactive Techniques},
+ series = {SIGGRAPH '88},
+ year = {1988},
+ isbn = {0-89791-275-6},
+ pages = {17--20},
+ numpages = {4},
+ url = {http://doi.acm.org/10.1145/54852.378457},
+ doi = {10.1145/54852.378457},
+ acmid = {378457},
+ publisher = {ACM},
+ address = {New York, NY, USA},
+ keywords = {linear edge function, parallel processing, polygon rasterization, sub-pixel vertices},
+}
+
+@article{K2018Optim,
+  journal = {Journal of Open Source Software},
+  doi = {10.21105/joss.00615},
+  issn = {2475-9066},
+  number = {24},
+  publisher = {The Open Journal},
+  title = {Optim: A mathematical optimization package for Julia},
+  url = {http://dx.doi.org/10.21105/joss.00615},
+  volume = {3},
+  author = {K Mogensen, Patrick and N Riseth, Asbjørn},
+  pages = {615},
+  date = {2018-04-04},
+  year = {2018},
+  month = {4},
+  day = {4},
 }
