### 2019-09-09 84e6340
Add citation information
