+++ README.md
@@ -122,14 +122,19 @@ $ wget https://raw.githubusercontent.com/tejank10/Duckietown.jl/master/src/meshe
 [Duckietown.jl](https://github.com/tejank10/Duckietown.jl) uses RayTracer.jl for generating renders
 of a self-driving car environment. For more complex examples of RayTracer, checkout that project.
 
-## JULIACON PAPER
+## SUPPORTING AND CITING:
 
-## Build Instructions
+This software was developed as part of academic research. If you would like to help support it, please star the repository. If you use this software as part of your research, teaching, or other activities, we would be grateful if you could cite:
 
-```bash
-$ cd paper
-$ latexmk -bibtex -pdf paper.tex
-$ latexmk -c
+```
+@misc{pal2019raytracerjl,
+    title={{RayTracer.jl: A Differentiable Renderer that supports Parameter Optimization for Scene Reconstruction}},
+    author={Avik Pal},
+    year={2019},
+    eprint={1907.07198},
+    archivePrefix={arXiv},
+    primaryClass={cs.GR}
+}
 ```
 
 ## CURRENT ROADMAP
