+++ README.md
@@ -14,21 +14,6 @@
 A Ray Tracer written completely in Julia. This allows us to leverage the AD capablities provided
 by Zygote to differentiate through the Ray Tracer.
 
-## SUPPORTING AND CITING:
-
-This software was developed as part of academic research. If you would like to help support it, please star the repository. If you use this software as part of your research, teaching, or other activities, we would be grateful if you could cite:
-
-```
-@misc{pal2019raytracerjl,
-    title={{RayTracer.jl: A Differentiable Renderer that supports Parameter Optimization for Scene Reconstruction}},
-    author={Avik Pal},
-    year={2019},
-    eprint={1907.07198},
-    archivePrefix={arXiv},
-    primaryClass={cs.GR}
-}
-```
-
 ## INSTALLATION
 
 The package is currently not registered. So open up a Julia 1.1+ repl and enter the pkg mode.
@@ -139,7 +124,7 @@ of a self-driving car environment. For more complex examples of RayTracer, check
 
 ## SUPPORTING AND CITING:
 
-This software was developed as part of academic research. If you would like to help support it, please star the repository. If you use this software as part of your research, teaching, or other activities, we would be grateful if you could cite:
+This software was developed as part of academic research. If you would like to help support it, please star the repository. If you use this software as part of your research, teaching, or other activities, we would be grateful if you could cite the following:
 
 ```
 @misc{pal2019raytracerjl,
