### 2019-09-09 d38c93f
Add citation information to the documentation as well
