+++ docs/src/index.md
@@ -30,6 +30,21 @@ image must be of reasonably high resolution.
 !!! note
     Only rendering is currently supported on GPUs. Gradient Computation is broken but
     will be supported in the future.
+    
+## Supporting and Citing
+
+This software was developed as part of academic research. If you would like to help support it, please star the repository. If you use this software as part of your research, teaching, or other activities, we would be grateful if you could cite:
+
+```
+@misc{pal2019raytracerjl,
+    title={{RayTracer.jl: A Differentiable Renderer that supports Parameter Optimization for Scene Reconstruction}},
+    author={Avik Pal},
+    year={2019},
+    eprint={1907.07198},
+    archivePrefix={arXiv},
+    primaryClass={cs.GR}
+}
+```
 
 ## Contribution Guidelines
 
