+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"#RayTracer-:-Differentiable-Ray-Tracing-in-Julia-1","page":"Home","title":"RayTracer : Differentiable Ray Tracing in Julia","text":"","category":"section"},{"location":"#","page":"Home","title":"Home","text":"RayTracer.jl is a library for differentiable ray tracing. It provides utilities for","category":"page"},{"location":"#","page":"Home","title":"Home","text":"Render complex 3D scenes.\nDifferentiate the Ray Tracer wrt arbitrary scene parameters for Gradient Based Inverse Rendering.","category":"page"},{"location":"#Installation-1","page":"Home","title":"Installation","text":"","category":"section"},{"location":"#","page":"Home","title":"Home","text":"Download Julia 1.1 or later. ","category":"page"},{"location":"#","page":"Home","title":"Home","text":"note: Note\nThis library won't work with Julia 1.0.","category":"page"},{"location":"#","page":"Home","title":"Home","text":"For the time being, the library is under active development and hence is not registered. But the master branch is pretty stable for experimentation. To install it simply open a julia REPL and  do ] add https://github.com/avik-pal/RayTracer.jl.","category":"page"},{"location":"#","page":"Home","title":"Home","text":"The master branch will do all computation on CPU. To try out the experimental GPU support do ] add https://github.com/avik-pal/RayTracer.jl#ap/gpu. To observe the potential performance gains of using GPU you will have to render scenes having more number of objects and the 2D image must be of reasonably high resolution.","category":"page"},{"location":"#","page":"Home","title":"Home","text":"note: Note\nOnly rendering is currently supported on GPUs. Gradient Computation is broken but will be supported in the future.","category":"page"},{"location":"#Contents-1","page":"Home","title":"Contents","text":"","category":"section"},{"location":"#","page":"Home","title":"Home","text":"Pages = [\"index.md\",\n         \"getting_started/rendering.md\",\n         \"getting_started/optimization.md\",\n         \"api.md\"]\nDepth = 3","category":"page"},{"location":"#Contributions-1","page":"Home","title":"Contributions","text":"","category":"section"},{"location":"#","page":"Home","title":"Home","text":"This package is written and maintained by Avik Pal. Please fork and send a pull request or create a GitHub issue for bug reports or feature requests.","category":"page"},{"location":"#Index-1","page":"Home","title":"Index","text":"","category":"section"},{"location":"#","page":"Home","title":"Home","text":"","category":"page"},{"location":"getting_started/rendering/#Rendering-1","page":"Rendering","title":"Rendering","text":"","category":"section"},{"location":"getting_started/rendering/#A-Step-By-Step-Introduction-to-Rendering-1","page":"Rendering","title":"A Step-By-Step Introduction to Rendering","text":"","category":"section"},{"location":"getting_started/rendering/#","page":"Rendering","title":"Rendering","text":"The very first step for in any application of the ray tracer is to define the scene parameters. This refers to","category":"page"},{"location":"getting_started/rendering/#","page":"Rendering","title":"Rendering","text":"Screen Size: We are defining it as a named tuple but it can be anything you wish.\nscreen_size = (w = 512, h = 512)\nLight Source(s): We can either define a single light source or a vector of light s
[…]
