+++ CITATION.bib
@@ -1,0 +1,15 @@
+@ARTICLE{2019arXiv190707198P,
+       author = {{Pal}, Avik},
+        title = "{RayTracer.jl: A Differentiable Renderer that supports Parameter Optimization for Scene Reconstruction}",
+      journal = {arXiv e-prints},
+     keywords = {Computer Science - Graphics, Computer Science - Computer Vision and Pattern Recognition},
+         year = "2019",
+        month = "Jul",
+          eid = {arXiv:1907.07198},
+        pages = {arXiv:1907.07198},
+archivePrefix = {arXiv},
+       eprint = {1907.07198},
+ primaryClass = {cs.GR},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019arXiv190707198P},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
