### 2019-09-19 avik-pal/RayTracer.jl#9 @dourouc05
The file `paper/paper.bbl` should not be present on the repo, as it is automatically generated from the BibTeX file (add it to the gitignore?).

For me, that's it :)!
