### 2019-09-23 #37 @matbesancon
@snoeyink, @dourouc05 still missing some check boxes. If they are considered good, tick them so that we can wrap this up
