### 2019-10-02 avik-pal/RayTracer.jl#9 @matbesancon
@avik-pal was everything taken into account from this issue?
