### 2019-10-02 avik-pal/RayTracer.jl#9 @avik-pal
Yes. The last revision had all the changes.
