### 2019-10-02 avik-pal/RayTracer.jl#9 @matbesancon
Ok, I think you can close it then.
