### 2019-10-21 #37 @dourouc05
@snoeyink still has a few check boxes to tick, I think.
