### 2019-10-22 f03c297
Add missing DOIs
