+++ paper/ref.bib
@@ -3,6 +3,7 @@
   author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
   journal={SIAM review},
   volume={59},
+  doi={10.1137/141000671},
   number={1},
   pages={65--98},
   year={2017},
@@ -16,6 +17,7 @@
     volume = {37},
     number = {6},
     pages = {222:1--222:11},
+    doi = {10.1145/3272127.3275109},
     year = {2018}
 }
 
@@ -142,6 +144,7 @@
   Title = {{CasADi} -- {A} software framework for nonlinear optimization
            and optimal control},
   Journal = {Mathematical Programming Computation},
+  doi = {10.1007/s12532-018-0139-4},
   Year = {In Press, 2018},
 }
 
