### 2019-10-22 #37 @avik-pal
@matbesancon I am not sure what is the issue here I added the DOIs in the latest commit [f03c297](https://github.com/avik-pal/RayTracer.jl/commit/f03c2974a87f5f05e34c2d8f53f736e5a6bd5d4f)
