+++ paper/ref.bib
@@ -1,24 +1,35 @@
 @article{bezanson2017julia,
-  title={Julia: A fresh approach to numerical computing},
-  author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
-  journal={SIAM review},
-  volume={59},
-  doi={10.1137/141000671},
-  number={1},
-  pages={65--98},
-  year={2017},
-  publisher={SIAM}
+author = {Bezanson, Jeff. and Edelman, Alan. and Karpinski, Stefan. and Shah, Viral B.},
+title = {{Julia: A Fresh Approach to Numerical Computing}},
+journal = {SIAM Review},
+volume = {59},
+number = {1},
+pages = {65-98},
+year = {2017},
+doi = {10.1137/141000671},
+URL = {https://doi.org/10.1137/141000671},
+eprint = {https://doi.org/10.1137/141000671}
 }
 
 @article{Li:2018:DMC,
-    title = {Differentiable {M}onte {C}arlo {R}ay {T}racing through {E}dge {S}ampling},
-    author = {Li, Tzu-Mao and Aittala, Miika and Durand, Fr{\'e}do and Lehtinen, Jaakko},
-    journal = {ACM Trans. Graph. (Proc. SIGGRAPH Asia)},
-    volume = {37},
-    number = {6},
-    pages = {222:1--222:11},
-    doi = {10.1145/3272127.3275109},
-    year = {2018}
+ author = {Li, Tzu-Mao and Aittala, Miika and Durand, Fr{\'e}do and Lehtinen, Jaakko},
+ title = {{Differentiable Monte Carlo Ray Tracing Through Edge Sampling}},
+ journal = {ACM Trans. Graph.},
+ issue_date = {November 2018},
+ volume = {37},
+ number = {6},
+ month = dec,
+ year = {2018},
+ issn = {0730-0301},
+ pages = {222:1--222:11},
+ articleno = {222},
+ numpages = {11},
+ url = {http://doi.acm.org/10.1145/3272127.3275109},
+ doi = {10.1145/3272127.3275109},
+ acmid = {3275109},
+ publisher = {ACM},
+ address = {New York, NY, USA},
+ keywords = {differentiable programming, inverse rendering, ray tracing},
 }
 
 @article{kingma2014adam,
@@ -138,15 +149,25 @@
   bibsource = {dblp computer science bibliography, https://dblp.org}
 }
 
-@article{casadi,
-  Author = {Joel A E Andersson and Joris Gillis and Greg Horn
-            and James B Rawlings and Moritz Diehl},
-  Title = {{CasADi} -- {A} software framework for nonlinear optimization
-           and optimal control},
-  Journal = {Mathematical Programming Computation},
-  doi = {10.1007/s12532-018-0139-4},
-  Year = {In Press, 2018},
+@Article{casadi,
+author="Andersson, Joel A. E.
+and Gillis, Joris
+and Horn, Greg
+and Rawlings, James B.
+and Diehl, Moritz",
+title="CasADi: a software framework for nonlinear optimization and optimal control",
+journal="Mathematical Programming Computation",
+year="2019",
+month="Mar",
+day="01",
+volume="11",
+number="1",
+pages="1--36",
+issn="1867-2957",
+doi="10.1007/s12532-018-0139-4",
+url="https://doi.org/10.1007/s12532-018-0139-4"
 }
+
 
 @inproceedings{Appel:1968:TSM:1468075.1468082,
  author = {Appel, Arthur},
