### 2019-10-23 #37 @matbesancon
A few general comments:
Page 1, col 2, second paragraph, the claims in this paragraph are rather bold. "The AD software present in other languages" do you mean other than Julia or in statically typed languages? I think Tapenade is another reference for C and Fortran code.
Figure 2a, isn't the y-axis the Average time to render per pixel (time for each pixel would not be a scalar but a vector)
The link in footnote 1 is broken. Prefer links to the DOI with `examples/optim_compatibility.jl`. Same thing for footnote 3.

Page 3 section 5.1: we would have changed `Listing 1` -> Algorithm 1. Use either inline code or a listings environment to put the expression that follows.
