+++ paper/paper.tex
@@ -63,9 +63,7 @@ Ray Tracing is a technique in computer graphics for rendering 3D graphics with c
 
 Since ray tracing leverages the properties of the materials of the objects in the scene, a natural extension to the rendering problem would be to extract the exact properties of the materials, lighting, and so on, given an image of a scene. This task is known as inverse rendering. Calculating analytic gradients for every single parameter of the scene is a very tedious process and prone to errors. This has made it a difficult task to present a general gradient-based inverse rendering method. As such, there is only one framework in our knowledge, redner~\cite{Li:2018:DMC}, which has been able to do so by using analytic gradients. However, we bypass this problem by using AD. There have also been attempts at making rasterization differentiable~\cite{softraster}; however, this involves making changes in the core technique which is against our design principles.
 
-% Another major limitation that exists in the case of differentiability is the presence of geometric discontinuities, like the edge between two triangles. This makes the renderer non-differentiable at those points. Unfortunately, this is still one of the limitations (Section~\ref{sec:limitation}) of RayTracer.
-
-Rendering is a computationally expensive technique, and so it is generally done in static languages like C++. Developing software in such languages are incredibly time-consuming. Also, most languages lack the support of the state of the art automatic differentiation tools like Zygote~\cite{DBLP:journals/corr/abs-1810-07951}, Jax~\cite{jax}, which are generally implemented for high-level languages like Julia and Python. As such, it is challenging to develop differentiable renderers in those languages and then interface with popular deep learning software. The AD software present in other languages like CasADi~\cite{casadi} does not seamlessly integrate with packages, which means one needs to rewrite the software to use specific AD tools.
+Rendering is a computationally expensive technique, and so it is generally done in static languages like C++. Developing software in such languages are incredibly time-consuming. Also, most languages lack the support of the state of the art automatic differentiation tools like Zygote~\cite{DBLP:journals/corr/abs-1810-07951}, Jax~\cite{jax}, which are generally implemented for high-level languages like Julia and Python. As such, it is challenging to develop differentiable renderers in those languages and then interface with popular deep learning software. Most of the other existing AD softwares, which the authors are familiar with, like CasADi~\cite{casadi} does not seamlessly integrate with packages, which means one needs to rewrite the software to use specific AD tools.
 
 In this paper, we explore the idea of differentiability through a renderer, by leveraging the AD in Julia~\cite{bezanson2017julia}. We present a fully general renderer capable of handling complex scenes and able to differentiate through them. We do not rely on analytic gradients but use source-to-source AD to generate efficient gradient code in the backward pass. Our renderer contains very little code for integration with Zygote, and hence, in theory, we can plug in any other AD software written in Julia.
 
@@ -169,23 +167,6 @@ RayTracer.jl can be used to solve a variety of inverse graphics problems. Since 
     \label{fig:ad_better}
 \end{figure*}
 
-% \begin{algorithm}[!htb]
-%     \caption{}
-%     \label{alg:inv_render}
-%     \SetAlgoLined
-%     \KwResult{Optimized set of Camera Parameters}
-%     Initial Guess of Parameters\;
-%     \While{not \textbf{converged} or iter < max\_iter}{
-%         gs = gradient (x $\rightarrow$ mean\_squared\_loss (render\_image (x), target\_img), params)\;
-%         \For{param in params}{
-%             update! (optimizer, param, gs[param])\;
-%         }
-%         \If{loss < tolerance}{
-%             converged = True\;
-%         }
-%     }
-% \end{algorithm}
-
 \begin{algorithm}[!htb]
 \DontPrintSemicolon
[…]
