### 2019-10-23 254596b
Changes as per editor's comments
