### 2019-10-23 ab427c1
Changes as per editor's comments
