### 2019-10-23 #37 @avik-pal
@matbesancon Thank you for your comments. I have tried to address them in the latest revision. As for `Listing 1 -> Algorithm 1`, this is incorrect, I am referring to the listing on page 2.
