### 2019-11-07 #37 @matbesancon
@avik-pal you have a weird annotation in the abstract, can you try to add the cls to your gitignore and recompile?
