### 2019-11-07 #37 @matbesancon
Algortihm 1, line 10, what is "optimizer" here?
