### 2019-11-12 #37 @avik-pal
> Algortihm 1, line 10, what is "optimizer" here?

It refers to the current state of the optimizer. Can be any first order optimizer like Adam, SGD.
