### 2019-11-12 #37 @matbesancon
It should be specified as an input / parameter of the algorithm
