+++ Project.toml
@@ -12,6 +12,14 @@ Images = "916415d5-f1e6-5110-898d-aaa5f9f070e0"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
+[compat]
+julia = "1"
+Compat = "3"
+Flux = "0.9"
+ImageMagick = "0"
+Images = "0"
+Zygote = "0.4"
+
 [extras]
 Distributed = "8ba89e20-285c-5b6f-9357-94700520ee1b"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
