+++ Project.toml
@@ -14,7 +14,7 @@ Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [compat]
 julia = "1"
-Compat = "3"
+Compat = "2"
 Flux = "0.9"
 ImageMagick = "0"
 Images = "0"
