### 2019-11-22 b7d8d3b
Update compat entries and dependenc versions
