### 2019-11-22 9578682
Update compat entries and dependenc versions
