+++ Project.toml
@@ -13,12 +13,12 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [compat]
-julia = "1"
 Compat = "2"
 Flux = "0.9"
 ImageMagick = "0"
 Images = "0"
 Zygote = "0.4"
+julia = "1"
 
 [extras]
 Distributed = "8ba89e20-285c-5b6f-9357-94700520ee1b"
