+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"#RayTracer-:-Differentiable-Ray-Tracing-in-Julia-1","page":"Home","title":"RayTracer : Differentiable Ray Tracing in Julia","text":"","category":"section"},{"location":"#","page":"Home","title":"Home","text":"<p align=\"center\">\n    <video width=\"512\" height=\"320\" autoplay loop>\n        <source src=\"./assets/udem1.webm\" type=\"video/webm\">\n    </video> \n</p>","category":"page"},{"location":"#","page":"Home","title":"Home","text":"RayTracer.jl is a library for differentiable ray tracing. It provides utilities for","category":"page"},{"location":"#","page":"Home","title":"Home","text":"Render complex 3D scenes.\nDifferentiate the Ray Tracer wrt arbitrary scene parameters for Gradient Based Inverse Rendering.","category":"page"},{"location":"#Installation-1","page":"Home","title":"Installation","text":"","category":"section"},{"location":"#","page":"Home","title":"Home","text":"Download Julia 1.0 or later. ","category":"page"},{"location":"#","page":"Home","title":"Home","text":"For the time being, the library is under active development and hence is not registered. But the master branch is pretty stable for experimentation. To install it simply open a julia REPL and  do ] add https://github.com/avik-pal/RayTracer.jl.","category":"page"},{"location":"#","page":"Home","title":"Home","text":"The master branch will do all computation on CPU. To try out the experimental GPU support do ] add https://github.com/avik-pal/RayTracer.jl#ap/gpu. To observe the potential performance gains of using GPU you will have to render scenes having more number of objects and the 2D image must be of reasonably high resolution.","category":"page"},{"location":"#","page":"Home","title":"Home","text":"note: Note\nOnly rendering is currently supported on GPUs. Gradient Computation is broken but will be supported in the future.","category":"page"},{"location":"#Supporting-and-Citing-1","page":"Home","title":"Supporting and Citing","text":"","category":"section"},{"location":"#","page":"Home","title":"Home","text":"This software was developed as part of academic research. If you would like to help support it, please star the repository. If you use this software as part of your research, teaching, or other activities, we would be grateful if you could cite:","category":"page"},{"location":"#","page":"Home","title":"Home","text":"@misc{pal2019raytracerjl,\n    title={{RayTracer.jl: A Differentiable Renderer that supports Parameter Optimization for Scene Reconstruction}},\n    author={Avik Pal},\n    year={2019},\n    eprint={1907.07198},\n    archivePrefix={arXiv},\n    primaryClass={cs.GR}\n}","category":"page"},{"location":"#Contribution-Guidelines-1","page":"Home","title":"Contribution Guidelines","text":"","category":"section"},{"location":"#","page":"Home","title":"Home","text":"This package is written and maintained by Avik Pal. Please fork and send a pull request or create a GitHub issue for bug reports. If you are submitting a pull request make sure to follow the official Julia Style Guide and please use 4 spaces and NOT tabs.","category":"page"},{"location":"#Adding-a-new-feature-1","page":"Home","title":"Adding a new feature","text":"","category":"section"},{"location":"#","page":"Home","title":"Home","text":"For adding a new feature open a Github Issue first to discuss about it.\nPlease note that we try and avoid having many primitive objects. This might speed up  rendering in some rare cases (as most objects will end up being represented as Triangles) but is really painful to maintain in the future.\nIf you wish to add rendering algorith
[…]
