+++ .travis.yml
@@ -1,31 +1,0 @@
-# Documentation: http://docs.travis-ci.com/user/languages/julia/
-language: julia
-codecov: true
-
-os:
-  - linux
-  - osx
-
-julia:
-  - 1.1
-  - nightly
-
-matrix:
-  allow_failures:
-    - julia: nightly
-
-jobs:
-  include:
-    - stage: "Documentation"
-      julia: 1.1
-      os: linux
-      script:
-        - julia --project=docs/ --color=yes -e 'using Pkg;
-                                                Pkg.develop(PackageSpec(path=pwd()));
-                                                Pkg.instantiate()'
-        - julia --project=docs/ --color=yes --compile=min docs/make.jl
-      after_success: skip
-      
-script:
-  - julia --color=yes -e 'using Pkg; Pkg.activate();
-                          Pkg.instantiate(); Pkg.test(coverage=true)'
