### 2020-04-14 c7a8996
Minimum julia version to v1.3
