### 2020-04-14 6d84916
Update the gradient defination for Materials
