### 2020-04-14 a459085
avoid possible type mismatch in `specular_exponent` and `reflection` fields
