### 2020-04-14 df05ffc
Fixes to the examples
