+++ .github/workflows/CompatHelper.yml
@@ -2,17 +2,20 @@ name: CompatHelper
 
 on:
   schedule:
-    - cron: '00 * * * *'
+    - cron: '00 00 * * *'
 
 jobs:
   CompatHelper:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        julia-version: [1.2.0]
-        julia-arch: [x86]
+        julia-version: [1.3]
+        julia-arch: [x64]
         os: [ubuntu-latest]
     steps:
+      - uses: julia-actions/setup-julia@latest
+        with:
+          version: ${{ matrix.julia-version }}
       - name: Pkg.add("CompatHelper")
         run: julia -e 'using Pkg; Pkg.add("CompatHelper")'
       - name: CompatHelper.main()
