+++ .github/workflows/CompatHelper.yml
@@ -1,7 +1,8 @@
 name: CompatHelper
 
 on:
-  schedule:
+  - push
+  - schedule:
     - cron: '00 00 * * *'
 
 jobs:
