+++ .github/workflows/CompatHelper.yml
@@ -1,8 +1,8 @@
 name: CompatHelper
 
 on:
-  - push
-  - schedule:
+  push:
+  schedule:
     - cron: '00 00 * * *'
 
 jobs:
