### 2020-04-14 38e9b82
CompatHelper: add new compat entry for "Zygote" at version "0.4"
