### 2020-04-14 624cfb1
CompatHelper: add new compat entry for "Flux" at version "0.10"
