### 2020-04-14 51bfb09
CompatHelper: add new compat entry for "Compat" at version "3.8"
