### 2020-04-14 671cc6d
CompatHelper: add new compat entry for "ImageMagick" at version "1.1"
