### 2020-04-14 9dc84c0
CompatHelper: add new compat entry for "Images" at version "0.22"
