### 2020-04-14 3f32fc1
CompatHelper: add new compat entry for "Flux" at version "0.10" (#21)
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Avik Pal <avikpal@iitk.ac.in>
