### 2020-04-14 4fd4d57
CompatHelper: add new compat entry for "Compat" at version "3.8" (#22)
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Avik Pal <avikpal@iitk.ac.in>
