### 2020-04-14 cb634af
CompatHelper: add new compat entry for "ImageMagick" at version "… (#23)
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Avik Pal <avikpal@iitk.ac.in>
