### 2020-04-14 7a5607e
CompatHelper: add new compat entry for "Images" at version "0.22" (#24)
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Avik Pal <avikpal@iitk.ac.in>
