### 2020-04-15 d26363e
Update Manifest.toml docs
