+++ Project.toml
@@ -13,10 +13,10 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [compat]
-Images = "0.22"
 ImageMagick = "1.1"
 Compat = "3.8"
 Flux = "0.10"
+Images = "0.22"
 Zygote = "0.4"
 julia = "1"
 
