### 2020-04-15 5dcbec1
Update installation instructions
