+++ .github/workflows/pkg-update.yml
@@ -1,57 +1,0 @@
-name: pkg-update
-
-on:
-  schedule:
-    - cron: '0 0 * * *'
-  push:
-    branches:
-      - actions/trigger/pkg-update
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-
-      # https://github.com/tkf/julia-merge-except-manifests
-      - name: 'Merge master to create-pull-request/pkg-update'
-        uses: tkf/julia-merge-except-manifests@v1
-        with:
-          target: create-pull-request/pkg-update
-
-      # https://github.com/tkf/julia-update-manifests
-      - name: Update */Manifest.toml
-        uses: tkf/julia-update-manifests@v1
-        with:
-          version: '1.3'
-          projects: test/environments/main docs
-      - name: Update test/environments/jl10/Manifest.toml
-        uses: tkf/julia-update-manifests@v1
-        with:
-          version: '1.0'
-          projects: test/environments/jl10
-
-      # https://github.com/peter-evans/create-pull-request
-      # https://github.com/marketplace/actions/create-pull-request
-      - name: Create Pull Request
-        uses: peter-evans/create-pull-request@v1
-        with:
-          token: ${{ secrets.GITHUB_TOKEN }}
-          commit-message: Update */Manifest.toml
-          title: 'Update */Manifest.toml'
-          body: |
-            ## Commit Message
-            Update */Manifest.toml
-          # When merging PR via Mergify using squash method, it can
-          # create large commit message by listing titles of all
-          # commits (default behavior of GitHub squash merge).
-          # Overriding the message here by using Mergify's API.
-          #
-          # --- Commit Message and squash Method
-          # https://doc.mergify.io/actions.html#commit-message-and-squash-method
-          labels: no changelog
-          branch: create-pull-request/pkg-update
-          branch-suffix: none
-          base: master
-      - name: Check output environment variable
-        run: echo "Pull Request Number - ${{ env.PULL_REQUEST_NUMBER }}"
