+++ Project.toml
@@ -1,21 +1,22 @@
 name = "RayTracer"
 uuid = "60dacb86-48ff-11e9-0f01-03ab8794bbc9"
 authors = ["Avik Pal <avikpal@iitk.ac.in>"]
-version = "0.1.2"
+version = "0.1.3"
 
 [deps]
 Compat = "34da2185-b29b-5c13-b0c7-acf172513d20"
 Distributed = "8ba89e20-285c-5b6f-9357-94700520ee1b"
 Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
+ImageIO = "82e4d734-157c-48bb-816b-45c225c6df19"
 ImageMagick = "6218d12a-5da1-5696-b52f-db25d2ecc6d1"
 Images = "916415d5-f1e6-5110-898d-aaa5f9f070e0"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [compat]
-ImageMagick = "1.1"
 Compat = "3.8"
 Flux = "0.10"
+ImageMagick = "1.1"
 Images = "0.22"
 Zygote = "0.4"
 julia = "1"
