+++ .github/workflows/ci.yml
@@ -9,8 +9,8 @@ jobs:
     strategy:
       matrix:
         version:
-          - '1.3'
-          - '1.4'
+          - '1.0.5'
+          - '1.4.1'
           - 'nightly'
         os:
           - ubuntu-latest
