### 2020-05-09 3e08584
CompatHelper: add new compat entry for "ImageIO" at version "0.1" (#32)
