### 2020-05-25 3b48021
CompatHelper: bump compat for "ImageIO" to "0.2" (#33)
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
