### 2020-07-11 5ea4ca5
CompatHelper: bump compat for "Flux" to "0.11"
