+++ Project.toml
@@ -15,7 +15,7 @@ Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [compat]
 Compat = "3.8"
-Flux = "0.10"
+Flux = "0.10, 0.11"
 ImageIO = "0.1, 0.2"
 ImageMagick = "1.1"
 Images = "0.22"
