### 2020-07-15 1bb5e05
CompatHelper: bump compat for "ImageIO" to "0.3"
