+++ Project.toml
@@ -19,7 +19,7 @@ Flux = "0.10"
 ImageIO = "0.1, 0.2"
 ImageMagick = "1.1"
 Images = "0.22"
-Zygote = "0.4"
+Zygote = "0.4, 0.5"
 julia = "1"
 
 [extras]
