### 2020-07-16 a1cd1d4
Remove manifest (#37)
* Delete Manifest.toml

* Delete Manifest.toml

* Delete Manifest.toml
