### 2020-07-16 9ea22ef
Remove manifest (#37)
* Delete Manifest.toml

* Delete Manifest.toml

* Delete Manifest.toml
