+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"getting_started/optim_compatibility/#Optimizing-Scene-Parameters-using-Optim.jl-1","page":"Optimizing using Optim.jl","title":"Optimizing Scene Parameters using Optim.jl","text":"","category":"section"},{"location":"getting_started/optim_compatibility/#","page":"Optimizing using Optim.jl","title":"Optimizing using Optim.jl","text":"In this tutorial we will explore the exact same problem as demonstrated in Inverse Lighting Tutorial but this time we will use the Optimization Package Optim.jl. I would recommend going through a few of the tutorials on Optim before starting this one.","category":"page"},{"location":"getting_started/optim_compatibility/#","page":"Optimizing using Optim.jl","title":"Optimizing using Optim.jl","text":"If you have already read the previous tutorial, you can safely skip to Writing the Optimization Loop using Optim. The part previous to this is same as the previous tutorial.","category":"page"},{"location":"getting_started/optim_compatibility/#","page":"Optimizing using Optim.jl","title":"Optimizing using Optim.jl","text":"using RayTracer, Images, Zygote, Flux, Statistics, Optim","category":"page"},{"location":"getting_started/optim_compatibility/#Script-for-setting-up-the-Scene-1","page":"Optimizing using Optim.jl","title":"Script for setting up the Scene","text":"","category":"section"},{"location":"getting_started/optim_compatibility/#","page":"Optimizing using Optim.jl","title":"Optimizing using Optim.jl","text":"screen_size = (w = 64, h = 64)\n\nscene = load_obj(\"./tree.obj\")\n\ncam = Camera(\n    Vec3(0.0f0, 6.0f0, -10.0f0),\n    Vec3(0.0f0, 2.0f0,  0.0f0),\n    Vec3(0.0f0, 1.0f0,  0.0f0),\n    45.0f0,\n    0.5f0,\n    screen_size...\n)\n\norigin, direction = get_primary_rays(cam)\n\nfunction render(light, scene)\n    packed_image = raytrace(origin, direction, scene, light, origin, 2)\n    array_image = reshape(hcat(packed_image.x, packed_image.y, packed_image.z),\n                          (screen_size.w, screen_size.h, 3, 1))\n    return array_image\nend\n\nshowimg(img) = colorview(RGB, permutedims(img[:,:,:,1], (3,2,1)))\n\nlight_gt = PointLight(\n    Vec3(1.0f0, 1.0f0, 1.0f0),\n    20000.0f0,\n    Vec3(1.0f0, 10.0f0, -50.0f0)\n)\n\ntarget_img = render(light_gt, scene)\n\nshowimg(zeroonenorm(render(light_gt, scene)))\n\nlight_guess = PointLight(\n    Vec3(1.0f0, 1.0f0, 1.0f0),\n    1.0f0,\n    Vec3(-1.0f0, -10.0f0, -50.0f0)\n)\n\nshowimg(zeroonenorm(render(light_guess, scene)))","category":"page"},{"location":"getting_started/optim_compatibility/#Writing-the-Optimization-Loop-using-Optim-1","page":"Optimizing using Optim.jl","title":"Writing the Optimization Loop using Optim","text":"","category":"section"},{"location":"getting_started/optim_compatibility/#","page":"Optimizing using Optim.jl","title":"Optimizing using Optim.jl","text":"Since, there is no direct support of Optim (unlike for Flux) in RayTracer the interface might seem a bit ugly. This is mainly due to the way the two optimization packages work. Flux allows inplace operation and ideally even RayTracer prefers that. But Optim requires us to give the parameters as an AbstractArray.","category":"page"},{"location":"getting_starte
[…]
