### 2020-07-26 9e338a9
Update paper.tex
