### 2020-07-26 #37 @matbesancon
@avik-pal the missing DOI seems valid, could you add it to the references?
