### 2020-07-26 dc8c092
Add DOI for softraster
