### 2020-07-26 #37 @matbesancon
wonderful thank you. I still have a tiny bug on our side (a random string appearing), and this will be good
