### 2020-07-27 #37 @matbesancon
@arfon do you know why this "Proceedings of JuliaCon"  line is appearing below the abstract here?  It might need a build from scratch on the server?
