### 2020-07-27 #37 @arfon
> @arfon do you know why this "Proceedings of JuliaCon" line is appearing below the abstract here? It might need a build from scratch on the server?

I just rebooted Whedon (which re-initializes the filesystem) and tried recompiling but the same issue is there....
