### 2020-09-10 c4ebccc
Cleanup old code and add some simple batched functions
