+++ src/camera.jl
@@ -1,168 +1,27 @@
-export Camera, get_primary_rays
+function _get_look_at_transform_rotation(camera_positions::AbstractMatrix{T},  # 3 x N
+                                         look_at::AbstractMatrix{T},  # 3 x (N / 1)
+                                         up::AbstractMatrix{T}) where T  # 3 x (N / 1)
+    @assert size(camera_positions, 1) == size(look_at, 1) == size(up, 1) == 3
 
-# ------ #
-# Camera #
-# ------ #
+    nbatch = size(camera_positions, 2)
+    z_ = reshape(_normalize(look_at .- camera_position, 1), 3, 1, nbatch)
+    x_ = reshape(_normalize(_cross(up, z_), 1), 3, 1, nbatch)
+    y_ = reshape(_normalize(_cross(z_, x_), 1), 3, 1, nbatch)
 
-"""
-    FixedCameraParams
-
-The Parameters of the [`Camera`](@ref) which should not be updated in inverse rendering
-are wrapped into this object.
-
-### Fields:
-
-* `vup`    - Stores the World UP Vector (In most cases this is the Y-Axis Vec3(0.0, 1.0, 0.0))
-* `width`  - Width of the image to be rendered
-* `height` - Height of the image to be rendered
-"""
-struct FixedCameraParams{T} <: FixedParams
-    vup::Vec3{T}
-    width::Int
-    height::Int
+    return hcat(x_, y_, z_)  # 3 x 3 x N
 end
 
-Base.show(io::IO, fcp::FixedCameraParams) =
-    print(io, "    Fixed Parameters:\n        World UP - ", fcp.vup,
-          "\n        Screen Dimensions - ", fcp.height, " × ", fcp.width)
 
-Base.zero(fcp::FixedCameraParams) = FixedCameraParams(zero(fcp.vup), zero(fcp.width),
-                                                      zero(fcp.height))
-
-"""
-    Camera
-
-The Perspective View Camera Model
-
-### Fields:
-
-* `lookfrom`    - Position of the Camera in 3D World Space
-* `lookat`      - Point in the 3D World Space where the Camera is Pointing
-* `vfov`        - Field of View of the Camera
-* `focus`       - The focal length of the Camera
-* `fixedparams` - An instance of [`FixedCameraParams`](@ref)
-
-### Available Constructors:
-
-* `Camera(;lookfrom = Vec3(0.0f0), lookat = Vec3(0.0f0),
-           vup = Vec3(0.0f0, 1.0f0, 0.0f0), vfov = 90.0f0,
-           focus = 1.0f0, width = 128, height = 128)`
-* `Camera(lookfrom::Vec3{T}, lookat::Vec3{T}, vup::Vec3{T}, vfov::R,
-          focus::R, width::Int, height::Int) where {T<:AbstractArray, R<:Real}` 
-"""
-struct Camera{T}
-    lookfrom::Vec3{T}
-    lookat::Vec3{T}
-    vfov::T
-    focus::T
-    fixedparams::FixedCameraParams{T}
+function look_at_view_transform(eye::AbstractMatrix{T},
+                                look_at::AbstractMatrix{T},
+                                up::AbstractMatrix{T}) where T
+    camera_positions = eye .+ look_at
+    rot_mat = _get_look_at_transform_rotation(camera_positions, look_at, up)
+    translate = -batched_mul(permutedims(rot_mat, (2, 1, 3)),
+                             repeat(reshape(camera_positions,
+                                            size(camera_positions, 1),
+                                            size(camera_positions, 2),
+                                            1),
+                                    1, 1, size(rot_mat, 3)))[1, :, :]
[…]
