+++ src/light.jl
@@ -1,146 +1,157 @@
-export PointLight, DistantLight
-
-# ----- #
-# Light #
-# ----- #
-
-"""
-    Light
-
-All objects emitting light should be a subtype of this. To add a custom light
-source, two only two functions need to be defined.
-
-* `get_direction(l::MyCustomLightSource, pt::Vec3)` - `pt` is the point receiving the
-                                                      light from this source
-* `get_intensity(l::MyCustomLightSource, pt::Vec3, dist::Array)` - `dist` is the array
-                                                                   representing the distance
-                                                                   of the point from the light
-                                                                   source
-"""
-abstract type Light end
-
-"""
-    get_shading_info(l::Light, pt::Vec3)
-
-Returns the `direction` of light incident from the light source onto that
-Point and the intensity of that light ray. It computes these by internally
-calling the `get_direction` and `get_intensity` functions and hence this
-function never has to be modified for any `Light` subtype.
-
-### Arguments:
-
-* `l`  - The Object of the [`Light`](@ref) subtype
-* `pt` - The Point in 3D Space for which the shading information is queried
-"""
-function get_shading_info(l::Light, pt::Vec3)
-    dir = get_direction(l, pt)
-    dist = sqrt.(l2norm(dir))
-    intensity = get_intensity(l, pt, dist)
-    return normalize(dir), intensity
+function compute_diffuse_lighting(normals::AbstractArray{T, 3},
+                                  colors::AbstractArray{T, 2},
+                                  directions::AbstractArray{T, 2}) where T
+    # normals --> 3 x F x N
+    # colors --> 3 x (N / 1)
+    # directions --> 3 x (N / 1)
+    colors = reshape(colors, 3, 1, size(colors, 2))  # 3 x 1 x (N / 1)
+    directions = reshape(directions, 3, 1, size(directions, 2))  # 3 x 1 x (N / 1)
+    normals = _normalize(normals, 1)
+    directions = _normalize(directions, 1)
+    return compute_diffuse_lighting(normals, colors, directions)
 end
 
-"""
-    get_direction(l::Light, pt::Vec3)
-
-Returns the direction of light incident from the light source to `pt`.
-"""
-function get_direction(l::Light, pt::Vec3) end
-
-"""
-    get_intensity(l::Light, pt::Vec3, dist::Array)
-
-Computed the intensity of the light ray incident at `pt`.
-"""
-function get_intensity(l::Light, pt::Vec3, dist::AbstractArray) end
-
-# --------------- #
-# - Point Light - #
-# --------------- #
-
-"""
-    PointLight
-
-A source of light which emits light rays from a point in 3D space. The
-intensity for this light source diminishes as inverse square of the distance
-from the point.
-
-### Fields:
-
-* `color`     - Color of the Light Rays emitted from the source.
-* `intensity` - Intensity of the Light Source. This decreases as ``\\frac{1}{r^2}``
-                where ``r`` is the distance of the point from the light source.
-* `position`  - Location of the light source in 3D world space.
-
-### Available Constructors:
-
-* `PointLight(;color = Vec3(1.0f0), intensity::Real = 100.0f0, position = Vec3(0.0f0))`
-* `PointLight(color, intensity::Real, position)`
-"""
-struct PointLight{T<:AbstractArray} <: Light
-    color::Vec3{T}
-    intensity::T
-    position::Vec3{T}
+function compute_diffuse_lighting(normals::AbstractArray{T, 3},
+                                  colors::AbstractArray{T, 2},
+                                  directions::AbstractArray{T, 3}) where T
[…]
