### 2020-10-09 #37 @avik-pal
@matbesancon Do we know the reason behind the "Proceedings of JuliaCon" appearing below the abstract? Even in my local setup everything seems fine.
