### 2020-10-09 #37 @avik-pal
@matbesancon It seems fixed now. Somehow the stale paper.pdf file from the repository was not being overwritten and it had the "Proceedings of Juliacon" below the abstract.
