### 2020-10-10 #37 @arfon
> @matbesancon It seems fixed now. Somehow the stale paper.pdf file from the repository was not being overwritten and it had the "Proceedings of Juliacon" below the abstract.

Yeah, I meant to update this thread earlier - we made some underlying changes to Whedon this week to improve cache expiration on the service so it's possible that this has also fixed the issue here...
