### 2020-10-10 #37 @avik-pal
@arfon Actually I had to delete the old compiled pdf from the repository for this to work.
