### 2020-10-10 #37 @matbesancon
> https://doi.org/10.1109/iccv.2019.00780 is INVALID because of 'https://doi.org/' prefix

@avik-pal I think this is the last step before victory :)
