+++ paper/ref.bib
@@ -127,7 +127,7 @@ eprint = {https://doi.org/10.1137/141000671}
   eprint    = {1904.01786},
   timestamp = {Wed, 24 Apr 2019 12:21:25 +0200},
   biburl    = {https://dblp.org/rec/bib/journals/corr/abs-1904-01786},
-  doi       = {https://doi.org/10.1109/iccv.2019.00780},
+  doi       = {10.1109/iccv.2019.00780},
   bibsource = {dblp computer science bibliography, https://dblp.org}
 }
 
