### 2020-10-11 #37 @arfon
> @matbesancon It seems fixed now. Somehow the stale paper.pdf file from the repository was not being overwritten and it had the "Proceedings of Juliacon" below the abstract.

Ahh, good to know!
