+++ README.md
@@ -12,7 +12,7 @@
     </video>
 </p>
 
-> This package was written in the early days of Flux / Zygote. Both these packages have significantly improved over time. Unfortunately the current state of this package of has not been updated to reflect those improvements. However, I do plan to do a major overhaul of the package around Jan 2021 to integrate these improvements along with the direct use of Flux3D.jl to use more efficient kernels for mesh operations.
+> This package was written in the early days of Flux / Zygote. Both these packages have significantly improved over time. Unfortunately the current state of this package of has not been updated to reflect those improvements. It also seems that Zygote.jl is in maintainence mode and there will be a gradual transition to Diffractor.jl. Once that AD package is released and things settle down with Flux.jl and Flux3D.jl I will start working on a major overhaul.
 
 A Ray Tracer written completely in Julia. This allows us to leverage the AD capablities provided
 by Zygote to differentiate through the Ray Tracer.
