+++ README.md
@@ -12,10 +12,9 @@
     </video>
 </p>
 
-> This package was written in the early days of Flux / Zygote. Both these packages have significantly improved over time. Unfortunately the current state of this package of has not been updated to reflect those improvements. It also seems that Zygote.jl is in maintainence mode and there will be a gradual transition to Diffractor.jl. Once that AD package is released and things settle down with Flux.jl and Flux3D.jl I will start working on a major overhaul.
+> This package was written in the early days of Flux / Zygote. Both these packages have significantly improved over time. Unfortunately the current state of this package of has not been updated to reflect those improvements. It also seems that it might be better to gradually transition to AD systems like Diffractor (which will potentially have good support for mutations) / define the adjoints directly using ChainRules since Zygote will likely not be having these features (note there has not been an official announcement and the statement is based on some discussions in zulip forum)
 
-A Ray Tracer written completely in Julia. This allows us to leverage the AD capablities provided
-by Zygote to differentiate through the Ray Tracer.
+A Ray Tracer written completely in Julia. This allows us to leverage the AD capablities provided by Zygote to differentiate through the Ray Tracer.
 
 ## INSTALLATION
 
